nghttp2: separate libnghttp2 and nghttp2
nghttp2 applications will search for nghttp2 headers in the configured
include path; however, they are not yet installed as the package is
still in the build process. To workaround this with minimal work (for
this example project), split the packages into a libnghttp2 package and
nghttp2 package for building/deploying the headers/static library and
building/deploying the nghttp2 application respectfully.

Signed-off-by: James Knight <james.d.knight@live.com>