.github: rework skip-project logic
A previous commit attempted to skip a job based off a matrix-provided
value [1]; however, the `matrix` context is not available during the
`jobs.<job_id>.if` workflow key, so this never works.

This commit attempts another approach by utilizing a step to help
trigger a failure to avoid running other jobs. Less than ideal, but
unknown if there is another approach with the existing job design.

[1]: 94809ddfb5bbb9babb7f497862450349e9be26e8
[2]: https://docs.github.com/en/actions/learn-github-actions/contexts#context-availability

Signed-off-by: James Knight <james.d.knight@live.com>