.github: drop bzr example test due to provided container limits
Since releng-tool no longer supports Python 2.7, there is no common
GitHub image that includes a Python 3 version for releng-tool to run on
with a Python 2.7 install to support an bzr installation. There is no
plans to make a custom test environment for this (at this time), so
dropping the test for now.

Signed-off-by: James Knight <git@jdknight.me>